+++ Black Jack/readme.md
@@ -1,0 +1,6 @@
+## BlackJack.py
+---
+A program that when ran, lets the user play a simple game of black jack with standard rules.
+
+###
+--
